chore: rename + alias ssg-helpers (#3518)
