+++ .wip/changes.md
@@ -1,0 +1,9 @@
+# Changes in v11
+
+## `interop`-mode has been removed
+
+...
+
+## SSG Helpers
+
+`createSSGHelpers` were for v9 which has now been removed. the v10 equivalent `createProxySSGHelpers` have been renamed to `createSSGHelpers` now instead. `createProxySSGHelpers` is now deprecated but aliased to `createSSGHelpers` for backwards compatibility.
