+++ packages/react-query/src/ssg/index.ts
@@ -1,2 +1,10 @@
-export { createProxySSGHelpers } from './ssgProxy';
-export type { DecoratedProcedureSSGRecord } from './ssgProxy';
+import { createSSGHelpers } from './ssg';
+
+export {
+  createSSGHelpers,
+  /**
+   * @deprecated - use `createSSGHelpers` instead
+   */
+  createSSGHelpers as createProxySSGHelpers,
+};
+export type { DecoratedProcedureSSGRecord } from './ssg';
